+++ src/Tests/Functions.fs
@@ -108,13 +108,30 @@ let tests = testList "Functions tests" [
         Expect.equal secrets.[1].Value (ExpressionSecret sa.Key) "Incorrect secret value"
         Expect.sequenceEqual secrets.[1].Dependencies [ vaults.resourceId "testfuncvault"; storageAccounts.resourceId "teststorage" ] "Incorrect secret dependencies"
     }
-
+    
     test "Supports dotnet-isolated runtime" {
-        let f = functions { name "func"; use_runtime (FunctionsRuntime.DotNetIsolated) }
+        let f = functions { name "func"; use_runtime (FunctionsRuntime.DotNetIsolated); }
         let resources = (f :> IBuilder).BuildResources Location.WestEurope
         let site = resources.[3] :?> Web.Site
         let settings = Expect.wantSome site.AppSettings "AppSettings should be set"
         Expect.equal settings.["FUNCTIONS_WORKER_RUNTIME"] (LiteralSetting "dotnet-isolated") "Should use dotnet-isolated functions runtime"
+    }
+    
+    test "Sets LinuxFxVersion correctly for dotnet runtimes" {
+        let getLinuxFxVersion f = 
+          let resources = (f :> IBuilder).BuildResources Location.WestEurope
+          let site = resources.[3] :?> Web.Site
+          site.LinuxFxVersion
+          
+        let f = functions { name "func"; use_runtime (FunctionsRuntime.DotNet50Isolated); operating_system Linux }
+        Expect.equal (getLinuxFxVersion f) (Some "DOTNET-ISOLATED|5.0") "Should set linux fx runtime"
+          
+        let f = functions { name "func"; use_runtime (FunctionsRuntime.DotNet60Isolated); operating_system Linux }
+        Expect.equal (getLinuxFxVersion f) (Some "DOTNET-ISOLATED|6.0") "Should set linux fx runtime"
+          
+        let f = functions { name "func"; use_runtime (FunctionsRuntime.DotNetCore31); operating_system Linux }
+        Expect.equal (getLinuxFxVersion f) (Some "DOTNETCORE|3.1") "Should set linux fx runtime"
+
     }
 
     test "FunctionsApp supports adding slots" {